Fix description parsing for empty sections
The MT940 structured description parser currently outputs a PHP "notice" when a section in the structured description is completely empty (e.g. `...?20EREF+                      ?....`). Depending on the PHP configuration, this notice can cause the entire program to abort. This commit allows empty sections, which will be output with their respective key and an empty string as the value.